Use Apples CryptoKit when available & fall back to CommonCrypto
This makes sure that Apple devices use Apples newer CryptoKit library for generating a SHA256 hash for improved safety & performance.

Note that there was a need for double-checking with both `#if canImport(CryptoKit)` and `if #available` to ensure full compatibility with all platforms & devices. See also [here](https://forums.swift.org/t/conditionally-use-cryptokit/30138/3).
